requirements.txt
google-adk>=1.25.1
google-cloud-bigquery>=3.25.0
